moves hacky temporary functions into a separate file
moves local import down
