+++ modules/mapping-utils/src/resolveAggregations.js
@@ -1,76 +1,13 @@
 import getFields from 'graphql-fields';
-import { flattenDeep, isArray, zipObject } from 'lodash';
-import { CONSTANTS } from '@arranger/middleware';
 
 import {
   buildQuery,
   buildAggregations,
   flattenAggregations,
 } from '@arranger/middleware';
+import { resolveSetsInSqon } from './hackyTemporaryEsSetResolution';
 
 let toGraphqlField = (acc, [a, b]) => ({ ...acc, [a.replace(/\./g, '__')]: b });
-
-const resolveSetIdsFromEs = es => setId =>
-  es
-    .search({
-      index: CONSTANTS.ES_ARRANGER_SET_INDEX,
-      type: CONSTANTS.ES_ARRANGER_SET_TYPE,
-      body: {
-        query: {
-          bool: {
-            must: { match: { setId } },
-          },
-        },
-      },
-    })
-    .then(({ hits: { hits } }) =>
-      flattenDeep(hits.map(({ _source: { ids } }) => ids)),
-    );
-
-const getSetIdsFromSqon = ({ content } = {}) =>
-  (isArray(content)
-    ? flattenDeep(
-        content.reduce(
-          (acc, subSqon) => [...acc, ...getSetIdsFromSqon(subSqon)],
-          [],
-        ),
-      )
-    : isArray(content?.value)
-      ? content?.value.filter(value => value.indexOf('set_id:') === 0)
-      : [...(content?.value.indexOf?.('set_id:') === 0 ? [content.value] : [])]
-  ).map(setId => setId.replace('set_id:', ''));
-
-const injectIdsIntoSqon = ({ sqon, setIdsToValueMap }) => ({
-  ...sqon,
-  content: sqon.content.map(op => ({
-    ...op,
-    content: !isArray(op.content)
-      ? {
-          ...op.content,
-          value: isArray(op.content.value)
-            ? flattenDeep(
-                op.content.value.map(
-                  value => setIdsToValueMap[value] || op.content.value,
-                ),
-              )
-            : setIdsToValueMap[op.content.value] || op.content.value,
-        }
-      : injectIdsIntoSqon({ sqon: op, setIdsToValueMap }).content,
-  })),
-});
-
-const resolveSetsInSqon = ({ sqon, es }) => {
-  const setIds = getSetIdsFromSqon(sqon || {});
-  return setIds.length
-    ? Promise.all(setIds.map(resolveSetIdsFromEs(es))).then(searchResult => {
-        const setIdsToValueMap = zipObject(
-          setIds.map(id => `set_id:${id}`),
-          searchResult,
-        );
-        return injectIdsIntoSqon({ sqon, setIdsToValueMap });
-      })
-    : sqon;
-};
 
 export default type => async (
   obj,
